Factor out sytnax highlighting code to separate class
